projeto04/requirements.txt
python-dotenv
openai
groq
